Refactor: Adopt Generated Classes and Optimize Serialization Process
The idea of these tests, is to test how far we can depend on the generated code, how we can add things to make them easier to use, and at the end lead make decision for creating the Exoscale java SDK.

Key changes include:

- Discarding the custom 'Instance' class in preference for the OpenAPI-generated equivalent.
- Implementing Jackson's ObjectMapper for efficient serialization of objects to JSON format within ExoscaleService. New methods, `createInstanceRequestBodyWithJackson` and `createAntiAffinityGroupRequestBodyWithJackson`, serialize 'Instance' and 'AntiAffinityGroup' objects correspondingly.
- Removal of the GSON dependency, after the usage of the Jackson's ObjectMapper
- Adding Lombok's `@Getter` and `@Setter` annotations to the 'SecurityGroup', 'InstanceType', and 'Template' classes to provide required fields for instance creation directly, thus facilitating easier use and clearer code.

This initiative is part of the test, with objective to see how far generated classes can be relied on. The aim is to simplify usage and establish a foundation for the final Exoscale Java SDK using openAPI generator.